refactor(chat): extract chat_quota.py (daily/anon quota + message validation)
Verbatim move of FREE_DAILY_LIMIT_* constants, _check_daily_quota,
_anon_quota_key, get_anonymous_quota_used, _check_anonymous_quota,
_validate_message. Re-imported into app.services.chat so api-layer
imports and test patch targets are unchanged. 819 tests pass.
