+++ backend/app/services/chat.py
@@ -2,7 +2,7 @@ import asyncio
 import json
 import logging
 import time as _time
-from datetime import UTC, date, datetime
+from datetime import UTC, datetime
 
 import httpx
 from sqlalchemy import delete, func, select, update
@@ -27,6 +27,15 @@ from app.database import async_session
 from app.models.chat import ChatAttachment, ChatMessage, ChatSession
 from app.models.user import User
 from app.schemas.chat import ChatResponse, ChatSource
+from app.services.chat_quota import (  # noqa: F401
+    FREE_DAILY_LIMIT_ANONYMOUS,
+    FREE_DAILY_LIMIT_USER,
+    _anon_quota_key,
+    _check_anonymous_quota,
+    _check_daily_quota,
+    _validate_message,
+    get_anonymous_quota_used,
+)
 from app.services.chat_trust import build_trust_status, persist_answer_diagnostic
 from app.services.citation_guard import enforce_citation_whitelist, log_mutations
 
@@ -46,9 +55,6 @@ from app.services.rag_retrieval import retrieve_rag_context
 
 logger = logging.getLogger(__name__)
 
-# Free daily limits for users without their own API key
-FREE_DAILY_LIMIT_USER = 200      # Logged-in users — effectively unlimited for normal use, caps abuse
-FREE_DAILY_LIMIT_ANONYMOUS = 10  # Anonymous users (encourage registration)
 
 # SSE keepalive cadence during LLM streaming. Sized for the tightest
 # documented idle-cut on the path: Cloudflare free-plan edge ≈100s,
@@ -656,88 +662,6 @@ def _resolve_fallback_llm_config() -> tuple[str, str, str, str] | None:
             provider = p
             break
     return url, settings.llm_fallback_api_key, model, provider
-
-
-async def _check_daily_quota(db: AsyncSession, user: User) -> None:
-    """Check and increment daily free chat quota. Raises QuotaExceededError if exceeded.
-
-    The increment runs as an **explicit UPDATE** rather than mutating
-    ``user`` attributes because ``user`` is loaded by ``get_optional_user``
-    on a *different* session from the one threaded into the streaming
-    chat path (see send_message_stream's prep-phase session). A
-    ``user.attr = value`` mutation against a session that doesn't own
-    the row gets silently dropped by ``flush()`` — no SQL is emitted,
-    quota stops incrementing, and free-tier limits stop applying.
-    The UPDATE-by-id form works regardless of which session loaded
-    ``user`` originally. The in-memory ``user`` is also patched so the
-    caller's view stays consistent within the same request.
-    """
-    today = date.today()
-    same_day = user.last_chat_date == today
-    current_count = user.daily_chat_count if same_day else 0
-    if current_count >= FREE_DAILY_LIMIT_USER:
-        raise QuotaExceededError(limit=FREE_DAILY_LIMIT_USER)
-    new_count = current_count + 1
-    await db.execute(
-        update(User)
-        .where(User.id == user.id)
-        .values(daily_chat_count=new_count, last_chat_date=today)
-    )
-    await db.flush()
-    # Keep the caller's in-memory User in sync with what we just wrote
-    # so any later attribute reads in the same request are coherent.
-    user.daily_chat_count = new_count
-    user.last_chat_date = today
-
-
-def _anon_quota_key(client_ip: str) -> str:
-    """Redis key for anonymous daily chat quota by IP."""
-    today = date.today().isoformat()
-    return f"chat:anon:{client_ip}:{today}"
-
-
-async def get_anonymous_quota_used(redis, client_ip: str) -> int:
-    """Get the number of chats used today by an anonymous IP."""
-    if not redis:
-        return 0
-    try:
-        val = await redis.get(_anon_quota_key(client_ip))
-        return int(val) if val else 0
-    except Exception:
-        return 0
-
-
-async def _check_anonymous_quota(redis, client_ip: str) -> None:
-    """Check and increment anonymous daily quota via Redis. Raises QuotaExceededError if exceeded."""
-    if not redis:
-        raise ServiceError("服务暂时不可用，请稍后重试")
-    key = _anon_quota_key(client_ip)
-    try:
-        current = await redis.incr(key)
-        if current == 1:
-            await redis.expire(key, 86400)  # 24h TTL
-        if current > FREE_DAILY_LIMIT_ANONYMOUS:
-            raise QuotaExceededError(limit=FREE_DAILY_LIMIT_ANONYMOUS)
-    except QuotaExceededError:
-        raise
-    except Exception:
-        logger.warning("Redis anonymous quota check failed", exc_info=True)
-
-
-def _validate_message(message: str) -> None:
-    """Validate chat message content.
-
-    The length cap must stay aligned with ``ChatRequest.message.max_length``
-    in app/schemas/chat.py — if the schema admits a longer message but
-    this service-layer check rejects it, the result is a stream-internal
-    ValidationError that surfaces in the UI as a generic
-    "请求失败，请重试" with no breadcrumb until you look at backend logs
-    (PR #651). Keep the two numbers in sync.
-    """
-    if not message or not message.strip():
-        raise ValidationError("消息不能为空")
-    if len(message) > 20000:
-        raise ValidationError("消息长度不能超过20000字")
 
 
 async def _resolve_session(
